requirements.txt
streamlit
pillow
numpy
opencv-python
tensorflow>=2.6
scikit-learn
matplotlib
imutils
